requirements.txt
SQLAlchemy~=2.0.44
fastapi~=0.118.3
pydantic~=2.12.0
PyYAML~=6.0.3
pytest~=9.0.1
httpx~=0.28.1
starlette~=0.48.0